refactor: extract shared learnings types into learnings-common.ts (#3477)
- Extract LearningType, LearningSource, validation functions, constants
- Apply skipStaleCheck pattern from memory-bridge-learning.ts
- Reduce code duplication between structured-learnings and memory-bridge-learning

Refs #3413
